Renaming examples. Joints no longer call get/set config tree.
