+++ examples/scene_baxter_pick_and_pass.ttt
