+++ examples/scene_youbot_navigation.ttt
